requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
pydantic
python-dotenv
